do not grab cache if it already exists or no target is found

+++ manuscripts/components/restore-cache.R
@@ -1,14 +1,19 @@
 #!/usr/bin/R
-
-library("utils")
-
-# restore cache file
+setwd("components")
 archive <- "template-manuscript-cache.tar.gz"
 
-setwd("components")
-download.file(paste0("http://two.ucdavis.edu/~cboettig/data/", archive), archive)
-untar(archive)
-unlink(archive)
+
+library("utils")
+library("httr")
+# restore cache file
+target <- paste0("http://two.ucdavis.edu/~cboettig/data/", archive)
+cache <- GET(target)
+if(cache$status_code != 404){
+  writeBin(content(cache, "raw"), archive)
+  untar(archive)
+  unlink(archive)
+}
+
 setwd("..")
